requirements.txt

# Core dependencies
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
# 核心深度学习框架
torch>=1.13.0
torchvision>=0.14.0



